Add Servo browser engine
This patch adds support for using Servo as an alternative browser
engine. Both Servo and WebKit are optional through their respective
cargo features, making it possible to build either, or both.

Currently this still uses a forked version of Servo, due to upstream's
lack of touch context menu support. However since two patches exist to
address this, we'll hopefully be able to switch to upstream soon.

One large advantage of the Servo engine is that it allows trying out
Kumo's UI without having to build the custom WebKit fork. It has also
been significantly easier to contribute to Servo compared to WPE WebKit.

The browser engine can be changed by opening the context menu and
selecting `Reload in WebKit/Servo`. To make using multiple engines
simultaneously convenient, this choice is cached whenever selected and
the next visit to the page will automatically load in the preferred
engine.

The following features are not implemented in Servo yet:
 - Downloads
 - Page Search
 - Tab audio playback indicator

+++ src/storage/history.rs
@@ -485,25 +485,23 @@ pub fn run_migrations(
     transaction: &Transaction<'_>,
     db_version: DbVersion,
 ) -> rusqlite::Result<()> {
-    match db_version {
-        // Create table if it doesn't exist yet.
-        DbVersion::Zero => {
-            let _ = transaction.execute(
-                "CREATE TABLE IF NOT EXISTS history (
+    // Create table if it doesn't exist yet.
+    if db_version == DbVersion::Zero {
+        let _ = transaction.execute(
+            "CREATE TABLE IF NOT EXISTS history (
                     uri TEXT NOT NULL PRIMARY KEY,
                     title TEXT DEFAULT '',
                     views INTEGER NOT NULL DEFAULT 1,
                     last_access INTEGER NOT NULL
                 )",
-                [],
-            )?;
-        },
-        // Delete all file/data URIs, since they were persisted incorrectly.
-        DbVersion::One => {
-            let _ = transaction.execute("DELETE FROM history WHERE uri LIKE 'file:%'", [])?;
-            let _ = transaction.execute("DELETE FROM history WHERE uri LIKE 'data:%'", [])?;
-        },
-        _ => (),
+            [],
+        )?;
+    }
+
+    // Delete all file/data URIs, since they were persisted incorrectly.
+    if db_version < DbVersion::Two {
+        let _ = transaction.execute("DELETE FROM history WHERE uri LIKE 'file:%'", [])?;
+        let _ = transaction.execute("DELETE FROM history WHERE uri LIKE 'data:%'", [])?;
     }
 
     Ok(())
